fixes yaml unmarshaling
